# docs/openAPI_document/streamflix_openapi.yaml
openapi: 3.1.0
info:
  title: StreamFlix API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /profiles/{id}:
    get:
      tags:
        - profiles
      summary: Get profile by ID
      description: Get specific profile. Supports JSON, XML, CSV.
      operationId: getProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - profile belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - profiles
      summary: Update profile
      description: Update a profile
      operationId: updateProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - profile belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - profiles
      summary: Delete profile
      description: Delete a profile
      operationId: deleteProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Profile deleted successfully
        '401':
          description: Unauthorized - JWT token missing or invalid
        '403':
          description: Forbidden - profile belongs to another user
        '404':
          description: Profile not found
...
  /profiles:
    get:
      tags:
        - profiles
      summary: Get profiles
      description: Get all profiles for the authenticated user. Supports JSON, XML, CSV.
      operationId: getProfiles
      responses:
        '200':
          description: Profiles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - profiles
      summary: Create profile
      description: Create a new profile for the authenticated user
      operationId: createProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
        required: true
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user by id
      description: Retrieve a user by id. Supports JSON, XML, and CSV formats via Accept header.
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - users
      summary: Update user
      description: Update a user's name or email. Users can only update their own account; administrators can update any account.
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
    UpdateUserRequest:
      properties:
        name:
          maxLength: 100
          type: string
        email:
          type: string
    UserResponse:
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - JUNIOR_EMPLOYEE
            - MID_EMPLOYEE
            - SENIOR_EMPLOYEE
            - USER
            - ADMINISTRATOR
        deleted:
          type: boolean
...
    UpdateProfileRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        avatar:
          type: string
          maxLength: 500
          minLength: 0
        age:
          type: integer
          format: int32
          maximum: 120
          minimum: 0
        contentFilters:
          type: array
          items:
            type: string
            enum:
              - VIOLENCE
              - FEAR
              - COARSE_LANGUAGE
              - NUDITY
              - DRUG_USE
              - SEXUAL_CONTENT
      required:
        - age
        - name
    ProfileResponse:
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        age:
          type: integer
          format: int32
        maturityLevel:
          type: string
        contentFilters:
          type: array
          items:
            type: string
            enum:
              - VIOLENCE
              - FEAR
              - COARSE_LANGUAGE
              - NUDITY
              - DRUG_USE
              - SEXUAL_CONTENT
...
    CreateProfileRequest:
      properties:
        name:
          type: string
        avatar:
          type: string
        age:
          type: integer
          format: int32
          maximum: 95
          minimum: 3
        contentFilters:
          type: array
          items:
            type: string
            enum:
              - VIOLENCE
              - FEAR
              - COARSE_LANGUAGE
              - NUDITY
              - DRUG_USE
              - SEXUAL_CONTENT
      required:
        - age
        - name
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT